refactoring code in the booking-list file
