+++ src/booking-list/BookingList.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { BookingInfoItems } from 'src/booking-list/BookingInfoItems';
 import { DateItems } from 'src/booking-list/DateItems';
 
 export const BookingList = () => {
@@ -8,7 +7,7 @@ export const BookingList = () => {
       <div className={'bookingWrapper'}>
         <div className={'bookingHeader cellDecoration'}>Booking list</div>
         <DateItems />
-        <BookingInfoItems />
+        <li>booking rides</li>
       </div>
     </div>
   );
